move everything from api/sovereignVenue into api/venue
